requirements.txt
amqp==2.5.0
atomicwrites==1.3.0
attrs==19.1.0
billiard==3.6.0.0
celery==4.3.0
environs==4.2.0
factory-boy==2.12.0
Faker==1.0.7
fire==0.1.3
importlib-metadata==0.18
kombu==4.6.3
marshmallow==2.19.5
more-itertools==7.0.0
numpy==1.16.4
packaging==19.0
pluggy==0.12.0
py==1.8.0
pymilvus==0.1.7
pyparsing==2.4.0
pytest==4.6.3
python-dateutil==2.8.0
python-dotenv==0.10.3
pytz==2019.1
redis==3.2.1
six==1.12.0
SQLAlchemy==1.3.5
text-unidecode==1.2
thrift==0.11.0
vine==1.3.0
wcwidth==0.1.7
zipp==0.5.1
mysql-connector-python==8.0.16
